Add custom_request, remove duplication (#184)
* Introduce custom_request to support all verbs

I believe, it's not useful to support every HTTP verb defined
everywhere with a custom method. But it should be possible to use those
verbs, thus we should introduce a method to trigger requests with a
custom verb. Unfortunately, there is already a public request method, so
I had to choose a different name.

* Avoid parsing the URI twice [Fixes #104]

* Allow long blocks in spec & gemspec

RSpec is encouraging very long blocks, so I don't think there's a way
around them.

The block in the gemspec is just configuration.